requirements/base.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-dotenv
dj-database-url
psycopg2-binary
